:sparkles: #19 add public and private routes with updated structure
