+++ src/routes/index.ts
@@ -1,0 +1,1 @@
+export { default as routes } from "./routes";
